+++ snippets/components/displays/cards/SolutionCard.jsx
@@ -22,7 +22,7 @@
  *
  * @example
  * import { BadgeWrapper, IconBadgeWrapper } from '/snippets/components/wrappers/badges/Badges.jsx'
- * import { ScrollBox } from '/snippets/components/wrappers/containers/ScrollBox.jsx'
+ * import { ScrollBox } from '/snippets/components/wrappers/containers/Layout.jsx'
  * import { SocialLinks } from '/snippets/components/elements/links/Links.jsx'
  *
  * <SolutionCard