+++ v2/gateways/custom/composables/linux/linuxSupport.jsx
@@ -1,4 +1,4 @@
-import { DynamicTable } from '/snippets/components/wrappers/tables/Table.jsx'
+import { DynamicTable } from '/snippets/components/displays/tables/Table.jsx'
 import { CustomCardTitle } from '/snippets/components/elements/text/Text.jsx'
 
 import {